RND-146: Extracted utility files
