+++ app/widgets/common/deploymentsView/header/RunWorkflowModal.consts.ts
@@ -1,0 +1,1 @@
+export const fetchedWorkflowFields = ['name', 'parameters'] as const;
