+++ app/widgets/common/deploymentsView/header/RunWorkflowModal.types.ts
@@ -1,0 +1,13 @@
+import type { fetchedWorkflowFields } from './RunWorkflowModal.consts';
+import type { Workflow } from '../../executeWorkflow';
+
+export type FetchedWorkflow = Pick<Workflow, typeof fetchedWorkflowFields[number]>;
+
+export type SimplifiedWorkflowParameter = FetchedWorkflow['parameters'] & {
+    name: string;
+};
+
+export interface EnhancedWorkflow extends Omit<FetchedWorkflow, 'parameters'> {
+    disabled: boolean;
+    parameters: SimplifiedWorkflowParameter[];
+}
